backend/requirements-eval.txt
ragas>=0.4.3
datasets>=2.18.0
pandas>=2.1.0
httpx>=0.25.0
tqdm>=4.66.0
openai>=1.40.0
tiktoken>=0.7.0
python-dotenv>=1.0.0
